Deploy earlier the ClusterRoleBindings granting  access for Admin/ViewerKubeconfig credentials
